Casting and security fixes

+++ lwlog/src/details/os/time_point_impl.h
@@ -4,59 +4,59 @@
 
 namespace lwlog::details::os
 {
-	template<typename LocalTimePolicy>
-	time_point<LocalTimePolicy>::time_point()
-	{
-		#ifdef _WIN32
-			::SYSTEMTIME m_now;
+    template<typename LocalTimePolicy>
+    time_point<LocalTimePolicy>::time_point()
+    {
+        #ifdef _WIN32
+            ::SYSTEMTIME m_now;
 
-			::GetSystemTime(&m_now);
+            ::GetSystemTime(&m_now);
 
-			year 	= m_now.wYear;
-			month 	= m_now.wMonth;
-			weekday = m_now.wDayOfWeek;
-			day 	= m_now.wDay;
-			minute 	= m_now.wMinute;
-			second 	= m_now.wSecond;
+            year 	= static_cast<uint16_t>(m_now.wYear);
+            month 	= static_cast<uint8_t>(m_now.wMonth);
+            weekday = static_cast<uint8_t>(m_now.wDayOfWeek);
+            day 	= static_cast<uint8_t>(m_now.wDay);
+            minute 	= static_cast<uint8_t>(m_now.wMinute);
+            second 	= static_cast<uint8_t>(m_now.wSecond);
 
-			hour = datetime::handle_timezone<LocalTimePolicy>(static_cast<std::uint8_t>(m_now.wHour));
+            hour = datetime::handle_timezone<LocalTimePolicy>(static_cast<std::uint8_t>(m_now.wHour));
 
-			::FILETIME now_ft;
-			::GetSystemTimePreciseAsFileTime(&now_ft);
+            ::FILETIME now_ft;
+            ::GetSystemTimePreciseAsFileTime(&now_ft);
 
-			::ULARGE_INTEGER ticks_since_windows_epoch;
-			ticks_since_windows_epoch.LowPart = now_ft.dwLowDateTime;
-			ticks_since_windows_epoch.HighPart = now_ft.dwHighDateTime;
+            ::ULARGE_INTEGER ticks_since_windows_epoch;
+            ticks_since_windows_epoch.LowPart = now_ft.dwLowDateTime;
+            ticks_since_windows_epoch.HighPart = now_ft.dwHighDateTime;
 
-			const std::uint64_t ticks_since_last_second{ ticks_since_windows_epoch.QuadPart % 10'000'000 };
+            const std::uint64_t ticks_since_last_second{ ticks_since_windows_epoch.QuadPart % 10'000'000 };
 
-			millisecond = m_now.wMilliseconds;
-			microsecond = (ticks_since_last_second / 10) % 1'000'000;
-			nanosecond = (ticks_since_last_second * 100) % 1'000'000'000;
-		#else
-			std::chrono::system_clock::time_point m_now;
+            millisecond = static_cast<uint16_t>(m_now.wMilliseconds);
+            microsecond = static_cast<uint32_t>((ticks_since_last_second / 10) % 1'000'000);
+            nanosecond = static_cast<uint32_t>((ticks_since_last_second * 100) % 1'000'000'000);
+        #else
+            std::chrono::system_clock::time_point m_now;
 
-		    m_now = std::chrono::system_clock::now();
-			const std::time_t now_time_t{ std::chrono::system_clock::to_time_t(m_now) };
+            m_now = std::chrono::system_clock::now();
+            const std::time_t now_time_t{ std::chrono::system_clock::to_time_t(m_now) };
 
-			std::tm details;
-			gmtime_r(&now_time_t, &details);
+            std::tm details;
+            gmtime_r(&now_time_t, &details);
 
-			year 	= static_cast<std::uint16_t>(details.tm_year) + 1900;
-			month 	= static_cast<std::uint8_t>(details.tm_mon) + 1;
-			weekday = static_cast<std::uint8_t>(details.tm_wday);
-			day 	= static_cast<std::uint8_t>(details.tm_mday);
-			minute 	= static_cast<std::uint8_t>(details.tm_min);
-			second 	= static_cast<std::uint8_t>(details.tm_sec);
+            year 	= static_cast<std::uint16_t>(details.tm_year) + 1900;
+            month 	= static_cast<std::uint8_t>(details.tm_mon) + 1;
+            weekday = static_cast<std::uint8_t>(details.tm_wday);
+            day 	= static_cast<std::uint8_t>(details.tm_mday);
+            minute 	= static_cast<std::uint8_t>(details.tm_min);
+            second 	= static_cast<std::uint8_t>(details.tm_sec);
 
-			hour = datetime::handle_timezone<LocalTimePolicy>(static_cast<std::uint8_t>(details.tm_hour));
+            hour = datetime::handle_timezone<LocalTimePolicy>(static_cast<std::uint8_t>(details.tm_hour));
 
-			const auto nanoseconds_since_unix_epoch{ 
-				std::chrono::duration_cast<std::chrono::nanoseconds>(m_now.time_since_epoch()).count() };
+            const auto nanoseconds_since_unix_epoch{ 
+                std::chrono::duration_cast<std::chrono::nanoseconds>(m_now.time_since_epoch()).count() };
 
-			millisecond = (nanoseconds_since_unix_epoch / 1'000'000) % 1'000;
-			microsecond = (nanoseconds_since_unix_epoch / 1'000) % 1'000'000;
-			nanosecond = nanoseconds_since_unix_epoch % 1'000'000'000;
-		#endif
-	}
+            millisecond = (nanoseconds_since_unix_epoch / 1'000'000) % 1'000;
+            microsecond = (nanoseconds_since_unix_epoch / 1'000) % 1'000'000;
+            nanosecond = nanoseconds_since_unix_epoch % 1'000'000'000;
+        #endif
+    }
 }
